@@ -1,20 +1,20 @@
 name: battery_plus_platform_interface
 description: A common platform interface for the battery_plus plugin.
-version: 0.2.0
+version: 0.3.0-nullsafety.0
 homepage: https://plus.fluttercommunity.dev/
 repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-133.7.beta <3.0.0"
   flutter: ">=1.17.0 <2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.2.3
-  plugin_platform_interface: ^1.0.2
+  meta: ^1.3.0-nullsafety
+  plugin_platform_interface: ^1.1.0-nullsafety
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.9.2
+  pedantic: ^1.10.0-nullsafety
