@@ -28,5 +28,5 @@ dev_dependencies:
   pedantic: ^1.9.2
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: ">=2.12.0-133.7.beta <3.0.0"
   flutter: '>=1.20.0'
